+++ Madera/src/Controller/MaderaModuleController.php
@@ -20,9 +20,14 @@ class MaderaModuleController extends AbstractController
      */
     public function index(MaderaModuleRepository $maderaModuleRepository): Response
     {
-        return $this->render('madera_module/index.html.twig', [
-            'madera_modules' => $maderaModuleRepository->findAll(),
-        ]);
+        $user = $this->getUser();
+        if($user == null){
+            return $this->redirectToRoute('app_login');
+        }else {
+            return $this->render('madera_module/index.html.twig', [
+                'madera_modules' => $maderaModuleRepository->findAll(),
+            ]);
+        }
     }
 
     /**
@@ -30,22 +35,27 @@ class MaderaModuleController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $maderaModule = new MaderaModule();
-        $form = $this->createForm(MaderaModuleType::class, $maderaModule);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if($user == null){
+            return $this->redirectToRoute('app_login');
+        }else {
+            $maderaModule = new MaderaModule();
+            $form = $this->createForm(MaderaModuleType::class, $maderaModule);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($maderaModule);
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($maderaModule);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('madera_module_index');
+                return $this->redirectToRoute('madera_module_index');
+            }
+
+            return $this->render('madera_module/new.html.twig', [
+                'madera_module' => $maderaModule,
+                'form' => $form->createView(),
+            ]);
         }
-
-        return $this->render('madera_module/new.html.twig', [
-            'madera_module' => $maderaModule,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -53,9 +63,14 @@ class MaderaModuleController extends AbstractController
      */
     public function show(MaderaModule $maderaModule): Response
     {
-        return $this->render('madera_module/show.html.twig', [
-            'madera_module' => $maderaModule,
-        ]);
+        $user = $this->getUser();
+        if($user == null){
+            return $this->redirectToRoute('app_login');
+        }else {
+            return $this->render('madera_module/show.html.twig', [
+                'madera_module' => $maderaModule,
+            ]);
+        }
     }
 
     /**
@@ -63,19 +78,24 @@ class MaderaModuleController extends AbstractController
      */
     public function edit(Request $request, MaderaModule $maderaModule): Response
     {
-        $form = $this->createForm(MaderaModuleType::class, $maderaModule);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if($user == null){
+            return $this->redirectToRoute('app_login');
+        }else {
+            $form = $this->createForm(MaderaModuleType::class, $maderaModule);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('madera_module_index');
+                return $this->redirectToRoute('madera_module_index');
+            }
+
+            return $this->render('madera_module/edit.html.twig', [
+                'madera_module' => $maderaModule,
+                'form' => $form->createView(),
+            ]);
         }
-
-        return $this->render('madera_module/edit.html.twig', [
-            'madera_module' => $maderaModule,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -83,12 +103,17 @@ class MaderaModuleController extends AbstractController
      */
     public function delete(Request $request, MaderaModule $maderaModule): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$maderaModule->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($maderaModule);
-            $entityManager->flush();
-        }
+        $user = $this->getUser();
+        if($user == null){
+            return $this->redirectToRoute('app_login');
+        }else {
+            if ($this->isCsrfTokenValid('delete' . $maderaModule->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($maderaModule);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('madera_module_index');
+            return $this->redirectToRoute('madera_module_index');
+        }
     }
 }
